requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.0
pyyaml==6.0.1

# Document loading
pymupdf==1.24.3
python-magic==0.4.27
markdown==3.6

# Chunking & embeddings
langchain-text-splitters==0.2.0
sentence-transformers==3.0.0

# Vector store
chromadb==0.5.0

# Hybrid search
rank_bm25==0.2.2

# LLM clients
ollama==0.2.1
anthropic>=0.30.0
google-generativeai==0.7.0

# HTTP (scraper)
httpx==0.27.0
beautifulsoup4==4.12.3
html2text==2024.2.26

# Evaluation
ragas==0.1.15
datasets==2.20.0

# Utilities
python-dotenv==1.0.1
cachetools==5.3.3
sse-starlette==2.1.0

# Redis cache
redis[asyncio]==5.0.8
